Add rudimentary toggle for fixture scope
The problem with the existing fixture "universe" in relation to the
functional tests is that:

* The functional tests need some of this state for testing the codelist
  builder: they need some codelist database to be available.
* The "universe" is generated once, when the tests are started (this
  happens via the top-level `conftest.py` importing everything from this
  `fixtures.py` module).
* The `live_server` fixture resets the database state after use in a
  test, and the functional tests require this fixture, so that an
  automated browser can interact with the server.

By switching the pytest fixture scope, we can force the fixtures to be
rebuilt. This is possibly considerably slower than creating specific
fixtures that we need, and using those, however. It does enable reuse of
much of the existing setup.

We could look to separate out the functional tests further and
provide those with their own fixtures for better isolation. This could
be considered in future, although one downside of that is a possible
duplication or near-duplication of some of the fixtures here.