Correct fixture from `icd10_data` to `snomedct_data`.
`test_matching_codes`, like all tests in this module, tests the `snomedct`
coding system and depends on it having been configured in a testing database.
While it passed when run alongside other tests in the module (which use the
correct `snomedct_data` fixture), it failed when run in isolation.

This issue appears to stem from a copy-paste error introduced during the
addition of this test alongside an ICD-10 test in commit e2188a4.