Allow multiple DB aliases in dynamic DB tests
This was originally done to facilitate making the remaining tests independent,
because while sketching this work out, it seemed required.

However, it turned out not to be the case.

That aside, it seems reasonable to consider retaining this change
because it subsequently makes some of the other database dependencies in
`coding_systems/base/tests/test_import_data_utils.py` explicit per test.

It also might make it in principle possible to reduce the repeated BNF
database and `CodingSystemRelease` setup in the BNF fixtures in
`coding_systems/base/tests/test_import_data_utils.py`, to a single
function that takes parameters. Even if we don't necessarily do that
now.