Implement basic multiple searches for testing
This refactors the code to make multiple searches at least possible.

As a bonus, all the tests now have the same structure, regardless of the
searches or their contents.

Validation of the final state is more complicated here.

In the simple test for two searches that is implemented, we explicitly
state every selection for the searches, which allows us to compute the
final included codelist.

This would allow testing of simple use cases, but does not yet allow for:

* computing the result for the "Searches" tab, because we would need to
  reapply the states of the final computed codelist back for each code
  in each search, if a later search overrides an earlier one
* more complex selections where we do not specify every included code in
  the tests and some are, for example, implicitly included; because we
  cannot easily compute the state without reimplementing the way the
  backend works in our test code (or, more realistically, using the
  backend code to derive the expected state)
* testing more esoteric behaviour (example: deleting searches, which may
  not be entirely well-defined behaviour yet in any case)
* and maybe more cases I have not yet thought of…