Pass latest published version rather than codelist
We want to display the number of codes in a codelist as part of the metadata. This information is tied to a particular version of the codelist as it can change between versions. The current pages (this_user and that_user) pass lists of CodelistVersions for draft and under review, but pass the Codelist for the published lists (both user and org lists). This commit changes it so that we now pass lists of CodelistVersions for each property. We previously just used the lists of Codelists to get their name and the url of the latest published version - by just passing the latest published version we can still get these.