Remove "Remove" link from `Search` card if not editable.
Builder pages are currently publicly visible, so that researchers can share
links of work-in-progress codelists. We expect this to continue. Therefore it's
possible to view a codelist builder page with searches that have a delete URL but
where the user doesn't have permission to delete -- this is exactly when
`isEditable` is `false`. We should not display the button if the user does not
have permission to delete the search.