Migration to update SNOMED CT human-readable name
The `choices` arg of the `CharField` definition for
Codelist.coding_system_id references the `CODING_SYSTEMS_CHOICES`
variable which iterates over all the available `CODING_SYSTEMS` and
returns a tuple of the coding system id and its human readable name.

f8b3e03 updated the human-readable name for the SNOMED CT coding system
to more accurately state that it is the UK Clinical Edition, which
therefore triggered a model change on Codelist which requires this
migration.