Add .env parameter for taking screenshots
The rationale is that we want the test to run everytime the functional tests are run so that we can fix the tests if a code change breaks them. However, we don't always want to generate screenshots because otherwise we end up committing a lot of image files. We set the variable to True in CI so the full tests do still run somewhere.