Implement VTM-VMP-AMP hierarchy
Define a hierarchy of entites from a subset of the dm+d coding_system
which will be presented to the builder via the `ancestor_relationships`
and `descendant_relationships` methods.

All AMPs have an ancestor VMP, most but not all VMPs have an ancestor
VTM so this must be handled in these methods.

AMPPs and VMPPs are excluded from this hierarchy as no user need for
them has been determined, and codes for these types of entity are not
present in the data sets made available through OpenSAFELY.
Additionally, AMPPs have dual ancestors of AMPs and VMPPs which would be
challenging to represent in these methods.