Improve codelist downloadability checking
Previously we relied on an assert in the formatted_table method of a
CodelistVersion to check if a codelist version is downloadable.

The version download view didn't account for this in all circumstances,
leading to unhandled AssertionErrors.

This commit removes this assert and instead makes both this view and
the codelist check API endpoint explicitly check the downloadability
of codelists versions and return a suitable error to users if not.