Cache version hierarchies as needed
The CodelistVersion model makes an assumption in the
`hierarchy` property that all versions in a
"builder compatible" coding system will already have cached hierarchies.

With dm+d going from a non-builder-compatible coding
system to a builder-compatible system, all old codelist
versions will not have cached hierarchies.

Unfortunately we can't import the `cache_hierarchy` method
from actions.py into models.py due to a circular import.
Therefore, we check for the presence of a cached hierarchy
in the version view and call it from there if not found.