Extend builder update to include metadata
We're going to want to update various bits of the metadata (description, methodology and references) from the builder. The easiest way to do that would seem to be to extend the `update` method. Currently that just takes any changes to the include/exclude, but now we also check if any of the metadata has been changed and update that as well

We also want to display the potential markdown fields in html, so I've refactored the render_markdown filter so it can be called here.