Check for minimum search length client-side
This adds a minimum search length for code and term search.
This is set for both for consistency, and making the code simpler.

However, for codes, this minimum length is unrestricted to be just
one character, since the original issue only specified a minimum
length for term searches: #2601.

The existing approach restricts input once that exceeds the maximum
length. This doesn't work well for minimum length, because a user
needs to start typing a search that is below the minimum allowed
 characters.

Note that `minLength` requires some user input before validating,
so this commit  also makes the search input a required attribute,
which also ensures that there is an input.