Validate codelist names are alphanumeric
The name field in the codelist create forms is
passed to one or more codelist creation methods in
actions.py.
These actions create handles for the new codelist.
Each handle has a slug, which is created by
via the django text utils slugify method.
This turns spaces into hyphens, then removes any
character that is not alphanumeric, hyphen, or
underscore.

Previously, users could create handles with
nonalphanumeric names, that in some cases would
slugify to an empty string.
This is an invalid slug, so we must prevent this
by ensuring the user-provided name will slugify
correctly.

This commit does this by adding a RegexValidator
to the Handle model, which allows only
alphanumeric characters, spaces, hyphens, and
underscores.

Both codelist creation forms (old and new style)
now have error handling logic for ValidationErrors
raised by this validator.

Handles can be owned by either a User or an
Organisation, so these fields can be blank.
An additional Check Constraint ensures that either
but not both are populated.
Previously, since no form field validation was
being called, the default `blank=False` model
rule was not being checked, but code in this
commit now runs those checks (along with the
aforementioned RegexValidator) so the rule for
these fields is changed to `blank=True` to
prevent spurious ValidationErrors.

Co-authored-by: KatieB5 <katie.bickford@thedatalab.org>