Validate codelist names are alphanumeric
The name field in the codelist create forms is
passed to one or more codelist creation methods in
actions.py.
These actions create handles for the new codelist.
Each handle has a slug, which is created by
via the django text utils slugify method.
This turns spaces into hyphens, then removes any
character that is not alphanumeric, hyphen, or
underscore.

Previously, users could create handles with
nonalphanumeric names, that in some cases would
slugify to an empty string.
This is an invalid slug, so we must prevent this
by ensuring the user-provided name will slugify
into something non-empty.

This commit does this by adding a RegexValidator
to the Handle model, which allows only
names with at least one sluggable character.
This is rather than requiring that the entire
field consist only of sluggable characters, and so
allows e.g., referring to Alzheimer’s or
Parkinson's with the standard punctuation.

Both codelist creation views (old and new style)
make use of the CodelistCreateForm so it is
convenient to handle the validation of user input
data here.
This is done by adding the model field's
validators to the corresponding form field in
`form_field_from_model()`.

Co-authored-by: KatieB5 <katie.bickford@thedatalab.org>
Co-authored-by: Mike Kelly <mike.kelly@thedatalab.org>