Move cached hierarchy checking to view wrapper
f299bfb added a check to see if a codelist version had a
cached_hierarchy and to build one if not to the codelist version view.

This was an adequate solution for the view in question, and if users
got to the csv download link via this page. However, going directly to
the csv download (as would be done by OpenSAFELY CLI for example) would
result in a 500 error.

This commit moves the check to the codelist version view wrapper such
that in all cases the cached_hierarchy check/build is run.