Add missing snomed_id header
Older dm+d codelists were uploaded with no checks.
When the dm+d coding system was updated to be "builder compatible",
the list of known headers was updated but this header value was omitted.

Checking for existing header values shows the only cases where there
are unknown headers (except for where there are missing headers, which
is accounted for) are "snomed_id" and a singular codelist with the
headers of ["x2", "x3"] which we can leave to fall to the default case.

```
import itertools
from codelists.models import CodelistVersion

dmd_codelistversions = CodelistVersion.objects.filter(
    codelist__coding_system_id="dmd")
unknown_headers = []
known_headers = [
    "ctv3id",
    "ctv3code",
    "ctv3_id",
    "snomedct_id",
    "snomedcode",
    "dmd",
    "dmd_id",
    "icd10",
    "icd code",
    "icd_code",
    "icd",
    "icd10_code",
    "id",
    "code",
]

for d in dmd_codelistversions:
    headers, *rows = d.table
    headers = [header.lower().strip() for header in headers]
    if not any([h for h in headers if h in known_headers]):
        unknown_headers.append(headers)

unknown_headers.sort()
list(h for h, _ in itertools.groupby(unknown_headers))
[
    ["", "obj_type", "vpid", "snomed_id", "bnf_code", "dmd_name"],
    [
        "0501011j0aaabab",
        "benzylpenicillin 600mg powder for solution for injection vials",
        "323398002",
        "vmp",
    ],
    ["319283006"],
    ["325278007"],
    ["35899811000001104", "amikacin 100mg/2ml solution for injection vials"],
    [
        "8093711000001103",
        "amp",
        "0601021a0aaaaaa",
        "glimepiride 2mg tablets (alliance healthcare (distribution) ltd)",
    ],
    ["obj_type", "vpid", "snomed_id", "bnf_code", "dmd_name", "bnf_name", "form_route"],
    [
        "vmp",
        "12816911000001103",
        "methotrexate 25mg/5ml oral solution",
        "1001030u0aaaaaa",
    ],
    ["vmp", "321987003", "citalopram 20mg tablets", "0403030d0aaaaaa"],
    ["vmp", "324430000", "trimethoprim 100mg tablets", "0501080w0aaadad"],
    ["x2", "x3"],
]
```