Pin minimum version of `requests` package
Dependabot want to use requests>=2.32.4 to fix a security vulnerability for both dev and prod, but can't bump it past 2.32.3. I'm not sure why because all the dependencies (django-anymail, and opentelemetry-exporter for prod at least) are pinned to compatible ranges (~=2.7 and >=2.4.3). Manually forcing it to use >=2.32.4 to get round this.