Tidy up orphaned codes when no longer included
Any orphaned codes, that are no longer included (explicitly or implicitly) can be removed. This prevents certain bad states occurring.

Also tidies up the comments for `get_codes_to_keep` as "descendants of included codes" and "codes with ancestor that is included" are two ways of saying the same thing.