Fix test affected by new orphan behaviour
Originally this test used the draft_with_no_searches. But we've changed it so that unselecting orphaned codes removes them from the codelist. The draft with no searches is entirely made up of orphaned codes, so to correctly test the logic in this test we instead need to use the draft with complete searches. The only difference is that the complete search one includes "439656005" while the no search one excludes it. So we tweak this test to use the complete search, but update the status of "439656005" and then the rest of the test works as previously.