Permit repeated conversion of BNF to dmd codelists
This updates the convert_bnf_codelist_version_to_dmd to account for
where a user requests a BNF codelist be converted to dm+d when there is
already such a codelist.

It previously threw an error, but now creates a new version of that
codelist with the latest conversion output and updates the metadata
of the destination codelist accordingly.

We have no way of checking what version of the mapping file was used
so it's hard to check whether a requested re-conversion is the same
as previously and so whether it's worth creating a new version.