Include dev dependencies in node-builder image
The node-builder image is used for linting and testing. Both of which use devDependencies. We change the image so that we don't need to use the `npx` commands which were starting to install the wrong versions.

The actual production build does not use node-builder, other than to copy the assets dir, so this is safe to do.