Use `str` path converter for `search_slug` in URL patterns.
This makes it clear that `search_slug` is not a slug according to
Django's concept of a slug. Add a comment to the action explaining some of the
reasons for that.

I'm not including tests as the `str` path converter is in-built Django
functionality and so broad that it's not clear what the test would actually do,
and the view doesn't actually use the slug so it's really non-critical.