Use `slug` path converter in URL patterns.
This makes it clear what kind of values may match in the pattern. The views
don't fail if a non-slug is usesd as they can perform their string-based
searches fine with a non-slug value. There is some value in using the path
converter for consistency, future-proofing, and to avoid URL pattern
collisions.

I'm not including tests as the `slug` path converter is in-built Django
functionality and the boilerplate call to each view with a non-valid slug to
see that the pattern is not matched doesn't seem valuable enough to merit the
extra test code.