Use `str` path converter for `tag_or_hash` in URL patterns.
`CodeliveVersion` hashes are hex-strings but tags are free text, so `tag_or_hash`
uses the less-restrictive `str:` path converter.

This makes it clear what kind of values may match in the pattern.  There is
some value in using a path converter here for consistency with other `urls.py`,
as a form of documentation, future-proofing, and to avoid URL pattern
collisions.

I'm not including tests as the `str:` path converter is in-built Django
functionality and so broad that it's not clear what the test would actually do.