Script to update missing cached_hierarchies
Various parts of the codebase make assumptions that if a CodelistVersion
is defined in a coding system that supports the creation of hierarchies
then it will have a cached_hierarchy.

This is not the case for CodelistVersions created before the hierarchy
creation methods were added to a coding system.
At the time of writing, this applies to the dm+d coding system.

cached_hierarchies are created on-demand in the CodelistVersion view
using the cached_hierarchy() method from codelists.actions, but it is
not possible to use this method in the coding system import routines
(which also make this false assumption) due to a circular import error.

This commit adds a script to update all missing cached_hierarchies for
all CodelistVersions in a given coding system.