Fix code derivation for headerless codelists
Old codelists uploaded via the "old" codelist create view did not have
and validation of the csv_data for the presence of a header.

Thus in some cases, there are codelists with no identifiable "code"
column header, and instead the first row is row data.

This commit fixes a bug whereby only the "code" column  of the first
row was re-added to the data rather than the full row, leading to an
indexing error.

CodelistVersion hashes are of their auto-incrementing integer id so
adding a new fixture causes a bump of ids and therefore hashes for all
fixtures following it.
Therefore, the expected hashes returned from the API have been updated.