Simplify Versions by reading value in component
Versions component is the only component which requires the versions
data loaded from the page.

Instead of passing the prop from the index, to the CodelistBuilder, to
the Sidebar, to the Versions component; we can instead read the
version data inside the Versions component.

As the versions data is only used in that component, we can move the
associated type definitions to the versions component file.