Apply `Codelist.name` validation in form not `codelist_create` action.
It seems like a good separation of concerns to have the form code validate the
user input, not the business-logic layer actions.py.

In order to avoid repeating the validator, it lives on the model field and is
inherited by the forms.

Also change the regex so that rather than requiring that the entire field
consist only of sluggable characters, we can relax the regex to require at
least one sluggable character. This means we can always generate a valid slug,
and allows e.g., referring to Alzheimer’s or Parkinson's with the standard
punctuation.