Add `sort_by_term` coding systems
Defaults to True, but set to False for ICD10 as we want to display those codes in code order rather than alphabetical. This was requested by a user and makes sense for "prefix" hierarchies where e.g. E10, E11, E12 are siblings and children of E1. Possibly should also set this for CTV3, but leaving it for now as it wasn't requested, and might be different to ICD10 in some way.