Extend automated screenshots to all docs images
This adds to the existing automated screenshots and generates all
documentation images.

In some cases, it's possible to use the existing code in
`test_build_codelist.py` and this works quite neatly.

However, for a lot of the screenshots, it's not easily
possibly without reworking that code. This is because the functions in
`test_build_codelist.py`, while allowing control of initial page navigation,
all assume that you want to actually perform an action. However, when
taking screenshots, we often want to pause and capture the screen before
then performing an action.

It might be possible to rewrite this test in future, and generalise the
code in `test_build_codelist.py` with that in mind. But initially having
the screenshots is an improvement on creating them by hand.