Update documentation after extensive builder work
This commit:

* switches to using Playwright generated screenshots
* updates the text to reflect those screenshots
* does some other clean up for brevity and clarity, and consistency of
  Markdown formatting

Note that while some text has been rewritten to use ordered lists, these
are incorrectly formatted as present (#2683). These are retained as
the kinds of instructions written this way are semantically ordered
lists. Hopefully the formatting gets fixed in future, but this is out of
scope for now.