Use AMP `desc` field as name, add `nm` synonyms
User feedback indicated strong preference for having the `descr`
field as the primary name for AMPs and the `nm` field as the secondary
name in the modal dialogue.

In all cases at the time of writing, the `descr` field is not null, but
if it were in a future release, having null names might cause
unexpected behaviour so this implementation coalesces `descr` with `nm`
just in case.

In many cases the `nm` field is a substring of the `descr` field but it
is included in case that is not always true.
This uses existing "synonym" mechanisms as it doesn't warrant one of its
own.