Correct use of `format_html()`
As written, this was ineffective.

`format_html()` takes a string to be formatted, and then escapes the
other arguments and keyword arguments passed to it. Passing only a
single, already formatted string results in no escaping.

Our incorrect use was deprecated in Django 5.0. It will result in a
`ValueError` in Django 6.0:

https://adamj.eu/tech/2023/06/15/format-html/