Add flags to force name and description changes
Previously passing a name and description was only honoured when creating a new codelist, but ignored if creating a new version of a codelist. For the PCD refsets we want to update the description, and possibly the name each time, so added flags in the bulk_import_codelists script to support this.