Add PCD refset update script
Gives us a way to reliably import the latest version of NHS PCD refsets. Due to previous errors manually importing codelists we have a few different scenarios to deal with:
1. A single codelist exists for the refset, and the slug is the refset id (e.g. ast_cod, flu_cod etc.)
2. A single codelist exists for the refset, and the slug is the slugified name (e.g. asthma-diagnosis-codes)
3. No codelists exist for the refset
4. Two codelists exist for the refset (both of the above)

We want all refsets to have the slug of the refset id from now on. Case 1 above is easy, we just create a new version. Case 2 is fine as well as we can match on name, then create a new handle for the codelist using the refset id - the old url will still work, but the slug is updated to the correct one. Case 3 is easy, we just create a new codelist. Case 4 is the tricky one, because we might not be able to update the name of the "correct" refset if it is an exact match (and therefore a clash) to the name of the "incorrect" refset. We can resolve that as a one-off by temporarily changing the names of all PCD refsets to include the suffix " (obsolete)" before running the script. As we force the name to be updated to the latest one, any matched codelists will lose the " (obsolete)" tag, just leaving ones that either have the wrong slug, or that don't exist in the latest refset zip - which should be marked as obsolete. We could manually go through these codelists and update their description to point to the real codelist if it exists.

Other changes to the json config:
- removed delimiter from json as it is ignored when the file is a ".csv" and a "," is assumed
- added ogl licence to auto-generated description