Fix CI file preventing log file output
The container seems to be given the name `docker-prod-1` instead of `docker_prod_1` so this command never gives the docker logs if the smoke test fails. Online discussion suggests its a platform issue, so given it might change, I've made the script robust to it by outputing the logs for all containers (which in practice is just the prod one).