Implement codelist cloning
Copy all codelist attributes apart from signoffs - the signatories
signed off the original, not this clone - and set the timestamps to now.

Status of the cloned version is set to UNDER_REVIEW as per the BNF-dmd
conversion process. We expect users to review the cloned codelist before
use.

Presume that the version the cloning user wants to clone is the most
recent one visible to them, rather than any older version or any newer
version(s) that they would not be able to see.

Deep copying all the related model objects is a faff, but has been done
in such a manner as to (hopefully) be resilient to new relations being
added rather than hard-coding those that exist at the time of writing.

Add codelist clone URL and view

Add clone button to template