Fix visibility of feedback-form
Previously, the <detials> tag of the
`_feedback_form.html` had an `open` attribute meaning the
feedback_form rendered in the open state by
default; this was helpful for development purposes
but for production we want it to render in the
closed state (i.e. as a button).

Previously, the feedback_form did not float over
the paragraph text in the page footer. A `z-index`
attribute was added to the `.feedback-form` CSS class
to correct this as we want the feedback_form element
to float over all other sections of the page.