Fix the API endpoint where no handles exist
This is associated with #2731.

It breaks because there's an assertion against there being a single `Handle`
for a `Codelist`. We currently have a case where there is a codelist
without a handle, though it's unclear why.

It would be preferable to ensure that codelists can't be created in a
broken state, but this works around the issue for now.