Handle duplicate username error
Sentry caught an unhandled ValidationError here. The username was ok when `form.is_valid()` was called (because it returned True), but then was a duplicate when the following line (`user=form.save()`) was called. I'm not sure why this happened (db issues, race condition?), and I'm tempted to not do anything, but at least now if it happens again the user will get the duplicate username warning and the error won't go to Sentry.