Limit number of codes returned by search
If there are too many codes returned we hit a limit in sqlite and an unhandled exception is thrown. This limit is currently 250,000 codes in an `IN` clause. However, large numbers of codes returned by a search is probably a sign the user has made a mistake or has not been specific enough. They would then need to include/exclude all codes which would take too long. So instead we set a much smaller limit (20,000 codes) and tell the user if their query exceeded that.