Allow local deployment testing
Add a just command (`just local-deploy`) to simulate deployment locally.

Full explanation of usage in deploy/local/LOCAL_DEPLOYMENT_TESTING.md

Updated the .dockerignore file so that when you build the image locally you don't get the database files bundled into the image. They're not meant to be built in to the image, and it can create enormous images if you have lots of local coding system databases.