Fix bug where metadata can't be edited
Move the QueryClient up to the CodelistBuilder level. In the MetadataTab component it was causing a bug where switching the tab caused the QueryClient to be recreated. This meant that changes to the description/methodology fields were persisted to the backend but not displayed locally.

By having it at this level it also means we can use the same one to keep the include/exclude state in sync with the backend.

Added a playwright test which failed before this change was committed to act as a regression test.