Fix typescript warnings
Rather than ignore the warnings we just correctly tell typescript what `data` might be.

Also drive by fix where an empty description returns as `<p></p>` was returning a blank display when it's better to show the `<thing> not provided` message. Added tests for that as well - and realised we need to run `queryClient.clear()` before each test to ensure the QueryClient state doesn't persist.