Pass the right methodology to the react app
The methodology can be changed during form initialization (where an empty string is converted back to the original template). In that situation we want the html display to be blank, but upon hitting Edit we want to display the original template. The existing implementation of passing `text` (to edit) and `html` (to display) makes this easy.