Improve style of rendered markdown
Users can put headings into their markdown, but the rendered html can then have h1, h2s etc which look out of place when underneath smaller headings. This reduces the heading size so that it looks better when viewed under the existing "Methodology" heading. The css needs setting twice at present, but once the "about" tab is refactored to a react app in a future commit this will not be necessary.