Add a context processor for taking screenshots
The feedback form button ends up in screenshots, often as slightly
awkward places (for example, in the middle of text), which may be due to
the way Playwright views the page.

Add a context processor so we can disable the button from appearing.
This has to be independent from `IN_PRODUCTION` since we use
`IN_PRODUCTION=True` in screenshots so that we don't get the indication bar
at the top page that tells us we're in development mode.

The alternative would be clicking to dismiss the button, or removing it
from the DOM somehow, but this feels cleaner than adding the button to
each test, and also makes the screenshot code more robust if the
feedback form button is changed or removed in future.