Allow cloning of Published or editable versions
The clone codelist action relies on the get_latest_visible_version
method from the Codelist model.
This returns the latest version for a Codelist that is _editable_ by a
user or has the Published status.

This commit changes the display logic for the codelist clone button such
that only the versions that meet the above "visible" criteria,
whilst removing a vestigial restriction that only user codelists are
cloneable.