Handle clone name collisions regardless of source
Previously we checked for name collisions only when a user was cloning
their own codelists.
This caused problems if a user cloned another user or organisation's
codelist that had the same name as one already in their codelists
(this was observered by a user re-cloning a codelist).

This change handles Handle collisions in all circumstances regardless
of who owns the source codelist.