Check both `name` and `slug` based uniqueness
The Handle model has UNIQUE constraints for both User+Name and
User+Slug (also for Organisation instead of User but doesn't apply here)
so we should check both of these.

We wouldn't expect Handles created as part of codelist cloning to fail
these constraint checks independently as the slug is created from the
name, but there may be edge cases with older data in production that we
are unaware of.