API to return any codelist with a `current` handle
This had been changed to filter out codelists without a handle, but by enforcing `handle_count=1` meant that codelists with multiple handles (a valid situation) were excluded. This change switches to use the `is_current` property of the handles to only return codelists with a current handle. This then correctly returns codelists with any number of handles as long as one of them is current.