Fail early if local db doesn't exist
When running locally it's possible for the coding_system db to not be available. This then causes all api requests to fail with 500 errors. This is a developer experience improvement that causes the script to fail early, and provide the user with the missing db alias so they can change it in their config, or download the actual one from the live server. No tests are added for this as it's complicated to make them pass in CI.