Exit early if first api call fails
If the first api POST fails, then it's very likely that there is an underlying issue that will cause all api calls to fail. Rather than printing hundreds of "failure" message, but no actual detail, this change causes the script to exit if the first api call fails and  also display the response body to help with debugging.