Check name and slug uniqueness for new clone names
When building a new unique handle for a cloned codelist,
we should check both name and slug as there are unique constraints
covering both of these fields.

The initial checking for entering the process to build a new unique
handle checked this, but the inner loop only checked name.

This commit adds this checking for slug uniqueness in that loop.

The slug for the newly-created codelists are created using the slugify
method on the name,
but previously-created codelists may have had their names altered by a
user which doesn't then re-slugify this name and keeps the old slug.

This led to slug collisions and unique constraint violation errors.