Better error if api lookup fails
The first thing the script does is to interrogate the latest releases from the api. This is not an authenticated end point so if it fails it means the site is down - most likely because the user is running locally and hasn't started the server, or because the host has been mistyped. This shows a much cleaner error than before.