Add NHSDrugRefset to the model
We will need this for correctly versioning the NHS drug refsets just like the existing NHS PCD refsets. They're likely identical models, but given they might change I've implemented as an abstract base class that they both inherit from so it is easy to add properties to one of them in future.