Refactor update_pcd_refsets
There are just 6 currently static things that need making variable in order to allow this script to also be used to load the drug refsets. This pulls those out to the top of the file so that when we make the change to support drug refsets it has a cleaner commit diff.