Evaluate exactly one `QuerySet` to fetch codelists.
Deleted codelists could cause an unhandled exception due to stale references in
the `all_published_codelists` iteration. Developers and users do sometimes delete
draft codelists.

`Handle.codelist` is a non-nullable foreign key, so it is never null. However
we evaluate the handles `QuerySet`, then iterate with a query per handle to
fetch its `codelist` based on the key in the in-memory instance. So the
codelist might not be in the database by the time that that query is evaluated
if it were deleted in that window. We know this page can take seconds to load.

Off-loading the codelist/version collection to the database in a single
`QuerySet` is the right thing to do anyway and would resolve this. And may help
with the [slowness of the
page](https://github.com/opensafely-core/opencodelists/issues/1640).

No new tests as I cannot think of a way to simulate a version being deleted
while the function runs without refactoring it in a weird way. We otherwise
expect the view to behave the same, just faster. Existing tests pass. Also
manually checked number of objects returned by default and with several search
terms is the same before and after the change.

From dev testing, with 2033 codelists on my local database.

Prior to this change accessing the full all published by organisation index
page executed 4779 queries in 700ms, overall CPU time about 2.3 seconds. I have
not analysed where all of these queries came from but there were at least: one
evaluating the handles QuerySet, one per handle.codelist access (as not
preselected) and one per has_published_versions() call as this evaluates
another QuerySet. So 2 x codelists + 1.

With the change there is only one query fetching all the handles and codelists,
taking 350ms, total time 380ms (so we also saved some Python execution time).
The SQL looks like...

```
(0.346) SELECT DISTINCT ... FROM "codelists_handle" INNER JOIN
    "codelists_codelist" ON ("codelists_handle"."codelist_id" =
    "codelists_codelist"."id") INNER JOIN "codelists_codelistversion" ON
    ("codelists_codelist"."id" = "codelists_codelistversion"."codelist_id") WHERE
    (NOT "codelists_codelist"."is_private" AND "codelists_handle"."is_current" AND
    "codelists_handle"."organisation_id" IS NOT NULL AND
    "codelists_codelistversion"."status" = 'published') ORDER BY
    "codelists_handle"."name" ASC;
```

Another 30 queries are executed when rendering the template, I think two per
codelist visible on the page. The template should not be doing this but that is
out of scope for this PR.

The under review version of the view for the OpenSAFELY organisation (eg
http://localhost:7000/codelist/opensafely/under-review/)  had 112 codelist
versions and executed 1308 queries in 400ms, about 700ms total CPU time. After
the change, that was 30 queries, 27 from the template, 3 from the function,
taking about 5ms, total time 290ms. Again, the templates should not execute
queries but that is out of scope.